[IMP] stock_storage_type: Ensure that a being emptied location can be filled in
If every possible destination location (e.g.: for pickable stock) is filled in with
products, but with one outgoing move is in progress (picked but not yet validated),
the location should become available to store incoming products.

The putaway strategy should propose that location. So, add a condition to select 'empty' and
'being emptied' locations along with the 'location_will_contain_product/lot_ids' property to
avoid too much restriction